+++ src/test/java/com/crud/modules/product/usecase/UpdateProductUnitTest.java
@@ -11,6 +11,8 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 
+import java.math.BigDecimal;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
@@ -29,7 +31,7 @@ class UpdateProductUnitTest {
   void UpdateProductSuccess() throws Exception {
     ProductRequest productRequest = new ProductRequest();
     productRequest.setQuantityStock(10);
-    productRequest.setPrice(250.00);
+    productRequest.setPrice(BigDecimal.valueOf(250));
     productRequest.setDescription("uni-Test");
     productRequest.setName("uni-Test");
